requirements.txt
Django==3.1.3
factory_boy==3.1.0
black==20.8b1
pre-commit==2.8.2
pre-commit-hooks==3.3.0
